(Course, Question_choices, Question) table's data has been added

+++ Tables/dbo.Question_Choices.sql
@@ -1,6 +1,6 @@
 CREATE TABLE [dbo].[Question_Choices]
 (
-[Question_ID] [float] NOT NULL,
+[Question_ID] [int] NOT NULL,
 [Choice] [nvarchar] (255) COLLATE SQL_Latin1_General_CP1_CI_AS NOT NULL
 ) ON [PRIMARY]
 GO
